Update (add webpack ang other configurations, add components)
