+++ .gitignore
@@ -12,4 +12,5 @@ Thumbs.db
 node_modules/
 
 # Папка с собранными файлами проекта
-build/
+public/bundle.js
+public/bundle.js.map
