+++ src/components/more.js
@@ -1,0 +1,2 @@
+export const createMoreTemplate = () =>
+  `<button class="load-more" type="button">load more</button>`;
