+++ webpack.config.js
@@ -1,14 +1,17 @@
-const path = require('path');
+const path = require(`path`);
 
 module.exports = {
-  mode: 'development',
-  entry: './src/main.js',
+  mode: `development`,
+  entry: `./src/main.js`,
   output: {
-    path: path.join(__dirname, 'public'),
-    filename: 'bundle.js'
+    path: path.join(__dirname, `public`),
+    filename: `bundle.js`
   },
-  devtool: 'source-map',
+  devtool: `source-map`,
   devServer: {
-    contentBase: path.join(__dirname, 'public')
+    contentBase: path.join(__dirname, `public`),
+    publicPath: `http://localhost:8081`,
+    compress: true,
+    watchContentBase: true
   }
 };
